clean up frontend some more, organize wagtail model, add rules
