+++ Pipfile
@@ -14,6 +14,7 @@ jinja2 = "*"
 django-user-agents = "*"
 wagtail = "*"
 gunicorn = "*"
+django-admin-view-permission = "*"
 
 [dev-packages]
