+++ app/models/__init__.py
@@ -5,4 +5,4 @@ from .signup_invite import SignupInvite
 from .supply_code import SupplyCode
 from .tag import Tag
 from .user import User
-from .wagtail_pages import *
+from .wagtail import *
