+++ app/models/wagtail/__init__.py
@@ -1,0 +1,3 @@
+from .root import *
+from .game_info import *
+from .news import *
