+++ app/static/stylesheets/components/_sidebar.scss
@@ -11,31 +11,6 @@
   border-right: 1px solid;
 }
 
-.ui-sidebar-nav {
-  padding: spacing(tight) 0 0 spacing(tight);
-  list-style: none;
-
-  li {
-    display: block;
-    list-style: none;
-  }
-
-  a {
-    padding: 0 spacing(tight);
-    margin-bottom: spacing(tighter);
-    display: block;
-    line-height: 1.25;
-    text-decoration: none;
-  }
-}
-
-.ui-sidebar-nav__title {
-  margin-bottom: spacing(tighter);
-  font-size: 12px;
-  font-weight: 500;
-  text-transform: uppercase;
-}
-
 
 .th--light {
   .ui-sidebar {
