+++ app/static/stylesheets/global.scss
@@ -31,6 +31,7 @@
 @import 'components/label';
 @import 'components/layout';
 @import 'components/logo';
+@import 'components/menu';
 @import 'components/sidebar';
 @import 'components/text-container';
 @import 'components/titlebar';
