+++ app/templates/jinja2/dashboard/base.html
@@ -20,43 +20,78 @@
   </div>
 </header>
 
-<nav class="ui-sidebar">
-  <ul class="ui-sidebar-nav">
-    <li><a href="{{ url('dashboard') }}">My Dashboard</a></li>
-  </ul>
-  {% if game is defined and game.is_running %}
-    <ul class="ui-sidebar-nav">
-      <h1 class="ui-sidebar-nav__title">Game Actions</h1>
-      <li><a href="{{ url('player_info') }}">Player Dashboard</a></li>
-      <li><a href="/dashboard/game_info/">Game Info</a></li>
-      <li><a href="{{ url('player_list') }}">Player List</a></li>
-      {% if (player is defined and not player.is_human) or user.is_moderator %}
-        <li><a href="#">Zombie Tree</a></li>
-      {% endif %}
+<aside class="ui-sidebar" id="side-nav">
+  <nav class="ui-menu">
+    <ul class="ui-menu__section">
+      <li><a href="{{ url('dashboard') }}">Home</a></li>
     </ul>
+  </nav>
+  {% if game is defined and game.is_running %}
+    <nav class="ui-menu">
+      <header class="ui-menu__header">
+        <h1 class="ui-menu__title">Game Actions</h1>
+        <a class="ui-menu__icon" href="#player-actions" data-target="#player-actions" data-toggle="collapse" aria-expanded="true">
+          <span class="oi oi-plus"></span>
+          <span class="oi oi-minus"></span>
+        </a>
+      </header>
+      <ul class="ui-menu__section collapse show" data-parent="#side-nav" id="player-actions">
+        <li><a href="{{ url('player_info') }}">Game Dashboard</a></li>
+        <li><a href="/dashboard/game_info/">Info & News</a></li>
+        <li><a href="{{ static('files/hvz-rules-v1.3.pdf') }}">Rules</a></li>
+        <li><a href="{{ url('player_list') }}">Player List</a></li>
+        {% if (player is defined and not player.is_human) or user.is_moderator %}
+          <li><a href="#">Zombie Tree</a></li>
+        {% endif %}
+      </ul>
+    </nav>
   {% endif %}
   {% if user.is_volunteer %}
-  <ul class="ui-sidebar-nav">
-    <h1 class="ui-sidebar-nav__title">Volunteer Actions</h1>
-    <li><a href="{{ url('signup_player') }}">Signup Players</a></li>
-  </ul>
+    <nav class="ui-menu">
+      <header class="ui-menu__header">
+        <h1 class="ui-menu__title">Volunteer Actions</h1>
+        <a class="ui-menu__icon " href="#volunteer-actions" data-target="#volunteer-actions" data-toggle="collapse" aria-expanded="false">
+          <span class="oi oi-plus"></span>
+          <span class="oi oi-minus"></span>
+        </a>
+      </header>
+      <ul class="ui-menu__section collapse" data-parent="#side-nav" id="volunteer-actions">
+        <li><a href="{{ url('signup_player') }}">Signup Players</a></li>
+      </ul>
+    </nav>
   {% endif %}
   {% if user.is_moderator %}
-    <ul class="ui-sidebar-nav">
-      <h1 class="ui-sidebar-nav__title">Moderator Actions</h1>
-      <li><a href="{{ url('manage_players') }}">Manage Players</a></li>
-      <li><a href="#">Manage OZ Pool</a></li>
-      <li><a href="{{ url('generate_supply_codes') }}">Generate Supply Codes</a></li>
-      <li><a href="/cms">Post Announcements & Missions (Wagtail)</a></li>
-    </ul>
+    <nav class="ui-menu">
+      <header class="ui-menu__header">
+        <h1 class="ui-menu__title">Moderator Actions</h1>
+        <a class="ui-menu__icon " href="#mod-actions" data-target="#mod-actions" data-toggle="collapse" aria-expanded="false">
+          <span class="oi oi-plus"></span>
+          <span class="oi oi-minus"></span>
+        </a>
+      </header>
+      <ul class="ui-menu__section collapse" data-parent="#side-nav" id="mod-actions">
+        <li><a href="{{ url('manage_players') }}">Manage Players</a></li>
+        <li><a href="#">Manage OZ Pool</a></li>
+        <li><a href="{{ url('generate_supply_codes') }}">Create Supply Codes</a></li>
+        <li><a href="/cms">Write News Posts (Wagtail)</a></li>
+      </ul>
+    </nav>
   {% endif %}
   {% if user.is_staff %}
-    <ul class="ui-sidebar-nav">
-      <h1 class="ui-sidebar-nav__title">Webmaster Actions</h1>
-      <li><a href="{{ url('admin:index') }}">Site Management (Django Admin)</a></li>
-    </ul>
+    <nav class="ui-menu">
+      <header class="ui-menu__header">
+        <h1 class="ui-menu__title">Webmaster Actions</h1>
+        <a class="ui-menu__icon " href="#webmaster-actions" data-target="#webmaster-actions" data-toggle="collapse" aria-expanded="false">
+          <span class="oi oi-plus"></span>
+          <span class="oi oi-minus"></span>
+        </a>
+      </header>
+      <ul class="ui-menu__section collapse" data-parent="#side-nav" id="webmaster-actions">
+        <li><a href="{{ url('admin:index') }}">Site Management (Django Admin)</a></li>
+      </ul>
+    </nav>
   {% endif %}
-</nav>
+</aside>
 
 <main class="ui-layout ui-layout--dashboard">
   <section class="ui-layout__item">
