+++ app/templates/jinja2/dashboard/moderator/manage_players.html
@@ -34,9 +34,9 @@
     <table class="table table-striped">
       <thead>
       <tr>
-        <th scope="col">Location</th>
-        <th scope="col">Number of emails sent</th>
-        <th scope="col">Number of emails acted upon</th>
+        <th scope="col">Signup Location</th>
+        <th scope="col">Emails sent</th>
+        <th scope="col">Emails acted on</th>
       </tr>
       </thead>
       <tbody>
