+++ app/templates/jinja2/dashboard/player.html
@@ -4,9 +4,9 @@
 {% block title %}Dashboard - Player Info{% endblock %}
 
 {% block pagetitle %}
-  {% if game.is_running %}{{ game.name }} - {% endif %}Player Dashboard
+  {% if game.is_running %}{{ game.name }} - {% endif %}Game Information
 {% endblock %}
-{% block pagesubtitle %}My Player Information{% endblock %}
+{% block pagesubtitle %}My Player Dashboard{% endblock %}
 
 {% block subcontent %}
 
@@ -70,7 +70,9 @@
     </div>
   {% else %}
   <div class="ui-text-container">
-    <p>As a spectator, you cannot access any actions that affect the game (reporting stuns/tags, redeeming supply codes, etc.). However, please feel free to check out some of the other pages on the sidebar.</p>
+    {% call card('') %}
+      <p>As a spectator, you cannot access any actions that affect the game (reporting stuns/tags, redeeming supply codes, etc.). However, please feel free to check out some of the other pages on the sidebar.</p>
+    {% endcall %}
   </div>
   {% endif %}
 {% elif game is defined and game.is_active %}
