+++ app/templates/jinja2/dashboard/player_list.html
@@ -4,9 +4,9 @@
 {% block title %}Dashboard - Player List{% endblock %}
 
 {% block pagetitle %}
-  {% if game.is_running %}{{ game.name }} - {% endif %}Player List
+  {% if game.is_running %}{{ game.name }} - {% endif %}Game Information
 {% endblock %}
-{% block pagesubtitle %}View Players{% endblock %}
+{% block pagesubtitle %}View Player List{% endblock %}
 
 {% block subcontent %}
   {% call card('Player List') %}
