+++ app/templates/jinja2/index.html
@@ -7,9 +7,8 @@
   <a href="/" class="sitelogo">{{ svg('site-logo-dark') }}</a>
   <nav>
     <ul>
-      <li>
-        <a class="text-light" href="#signup-schedule">Signup Schedule</a>
-      </li>
+      <li><a class="text-light" href="#signup-schedule">Signup Schedule</a></li>
+      <li><a class="text-light" href="{{ static('files/hvz-rules-v1.3.pdf') }}">Rules</a></li>
     </ul>
   </nav>
   <div class="actions">
@@ -91,5 +90,5 @@
 {% endblock %}
 
 {% block scripts %}
-<script src="{{ static('scripts/countdown.js') }}"></script>
+  <script src="{{ static('scripts/countdown.js') }}"></script>
 {% endblock %}
